requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.4
certifi==2018.8.24
cffi==1.12.3
chardet==3.0.4
click==6.7
colorama==0.3.9
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.1
django-heroku==0.3.1
django-toolbelt==0.0.1
entrypoints==0.2.3
Flask==1.1.1
gunicorn==20.0.0
html5lib==1.0.1
hyperlink==17.1.1
idna==2.7
incremental==17.5.0
ipykernel==4.10.0
ipython==6.5.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
lxml==4.3.3
MarkupSafe==1.0
mistune==0.8.3
more-itertools==7.2.0
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.6.0
pandocfilters==1.4.2
parsel==1.5.1
parso==0.4.0
pathlib2==2.3.2
pickleshare==0.7.4
pluggy==0.11.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
psycopg2==2.8.4
py==1.8.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.4.0
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
PySocks==1.6.8
pytest==3.8.1
pytest-runner==5.1
python-dateutil==2.7.3
pytz==2019.1
pywin32==223
pywinpty==0.5.4
pyzmq==17.1.2
queuelib==1.5.0
requests==2.19.1
Scrapy==1.6.0
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.12.0
static3==0.7.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1.1
traitlets==4.3.2
Twisted==17.5.0
urllib3==1.23
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
whitenoise==4.1.4
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
zope.interface==4.5.0